拆分panel展示状态
